refactor: Reestructuraciones en la carpeta de servicios (pagos y reportes), además de compilación de assets.
