+++ .env.example
@@ -62,3 +62,5 @@ VITE_PUSHER_APP_CLUSTER="${PUSHER_APP_CLUSTER}"
 PLACETOPAY_TRANKEY=
 PLACETOPAY_LOGIN=
 PLACETOPAY_URL=
+
+BACK_URL_LOG_VIEWER=
