+++ config/log-viewer.php
@@ -46,7 +46,7 @@ return [
     |
     */
 
-    'back_to_system_url' => config('app.url', null),
+    'back_to_system_url' => env('BACK_URL_LOG_VIEWER', null),
 
     'back_to_system_label' => null, // Displayed by default: "Back to {{ app.name }}"
